savingaccountcontroller api version, summarize an finished

+++ ArtemisBank.Core.Application/DTOs/Account/AssignSavingsAccountDto.cs
@@ -5,6 +5,6 @@ namespace ArtemisBank.Core.Application.DTOs.Account
     public class AssignSavingsAccountDto
     {
         public string ClientId { get; set; } = string.Empty;
-        public int InitialBalance { get; set; }
+        public decimal InitialBalance { get; set; }
     }
 }
